PR remarks
- Do not call the default constructor
- clusterCount() returns a size_t instead of uint

+++ src/libs/antares/study/parts/renewable/container.h
@@ -109,7 +109,7 @@ public:
     std::vector<RenewableCluster*> clusters;
     //! How many clusters have we got ?
     // Only available from the solver
-    inline uint clusterCount() const
+    inline size_t clusterCount() const
     {
         return clusters.size();
     }
